dom/base/test/test_bug1730284.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Test for bug 1730284 (throttling of same-origin iframes)</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<style>
  iframe {
    width: 10px;
    height: 10px;
  }
  .display-none {
    display: none;
  }
  .vis-hidden {
    visibility: hidden
  }
  .transparent {
    opacity: 0;
  }
  .zero-size {
    width: 0;
    height: 0;
    border: 0;
  }
  .offscreen {
    position: absolute;
    top: 500%;
  }
  .scroller {
    height: 100px;
    overflow: auto;
  }
  .scroller-padding {
    height: 500px;
  }
</style>
<iframe class="visible"></iframe>
<iframe class="display-none" data-throttled-expected></iframe>
<iframe class="vis-hidden"></iframe>
<iframe class="transparent"></iframe>
<iframe class="zero-size"></iframe>
<div class="scroller">
  <div class="scroller-padding"></div>
  <iframe class="scrolled-out-of-view" data-throttled-expected></iframe>
</div>
<iframe class="offscreen" data-throttled-expected></iframe>
<iframe class="offscreen zero-size" data-throttled-expected></iframe>
<iframe class="offscreen vis-hidden" data-throttled-expected></iframe>
<iframe class="offscreen transparent" data-throttled-expected></iframe>
<script>
add_task(async function() {
  await SimpleTest.promiseFocus(window);
  is(SpecialPowers.DOMWindowUtils.effectivelyThrottlesFrameRequests, false, "Should not be throttling main page");

  for (let frame of document.querySelectorAll("iframe")) {
    let shouldThrottle = frame.getAttribute("data-throttled-expected") !== null;
    is(SpecialPowers.getDOMWindowUtils(frame.contentWindow).effectivelyThrottlesFrameRequests, shouldThrottle, frame.className);
  }
});
</script>
